+++ Source/Plugins/Statistics/StatisticsFilters/FindSizes.h
@@ -201,8 +201,8 @@ protected:
 
 private:
   DEFINE_DATAARRAY_VARIABLE(int32_t, FeatureIds)
-  DEFINE_DATAARRAY_VARIABLE(float, Volumes)
-  DEFINE_DATAARRAY_VARIABLE(float, EquivalentDiameters)
+  DEFINE_DATAARRAY_VARIABLE(double, Volumes)
+  DEFINE_DATAARRAY_VARIABLE(double, EquivalentDiameters)
   DEFINE_DATAARRAY_VARIABLE(int32_t, NumElements)
 
   FindSizes(const FindSizes&);      // Copy Constructor Not Implemented
